Add Supplier dataclass and SuppliersDB modules
