+++ Main_v22.py
@@ -18,7 +18,6 @@ File Hopper – dual-mode (GUI if Tkinter available, else CLI)
 
 import os, sys, json, csv, shutil, argparse, tempfile, datetime, threading, re
 from collections import defaultdict
-from dataclasses import dataclass, asdict
 from typing import List, Dict, Optional, Tuple, Any
 
 import pandas as pd
@@ -34,7 +33,6 @@ try:
 except Exception:
     REPORTLAB_OK = False
 
-SUPPLIERS_DB_FILE = "suppliers_db.json"
 MIAMI_PINK = "#FF77FF"
 DEFAULT_FOOTER_NOTE = (
     "Gelieve afwijkingen schriftelijk te bevestigen. "
@@ -53,277 +51,8 @@ from helpers import (
     _build_file_index,
     _unique_path,
 )
-
-
-# --------------------------- Supplier model & mapping ---------------------------
-
-@dataclass
-class Supplier:
-    supplier: str
-    description: Optional[str] = None
-    supplier_id: Optional[str] = None
-    adres_1: Optional[str] = None
-    adres_2: Optional[str] = None
-    postcode: Optional[str] = None
-    gemeente: Optional[str] = None
-    land: Optional[str] = None
-    btw: Optional[str] = None
-    contact_sales: Optional[str] = None
-    sales_email: Optional[str] = None
-    phone: Optional[str] = None
-    favorite: bool = False
-
-    @staticmethod
-    def from_any(d: dict) -> "Supplier":
-        """Normalize kolomnamen uit CSV/JSON naar canonical fields."""
-        key_map = {
-            # naam & beschrijving
-            "supplier": "supplier",
-            "leverancier": "supplier",
-            "supplier name": "supplier",
-            "naam": "supplier",
-            "description": "description",
-            "beschrijving": "description",
-            "omschrijving": "description",
-            "notes": "description",
-            # id
-            "supplier_id": "supplier_id",
-            "supplier id": "supplier_id",
-            "id": "supplier_id",
-            # adres 1
-            "adres_1": "adres_1",
-            "adres 1": "adres_1",
-            "adres1": "adres_1",
-            "address_1": "adres_1",
-            "address 1": "adres_1",
-            "address": "adres_1",
-            "adress_1": "adres_1",
-            "adress 1": "adres_1",
-            "adress1": "adres_1",
-            "straat": "adres_1",
-            # adres 2
-            "adres_2": "adres_2",
-            "adres 2": "adres_2",
-            "adres2": "adres_2",
-            "address_2": "adres_2",
-            "address 2": "adres_2",
-            "adress_2": "adres_2",
-            "adress 2": "adres_2",
-            "adress2": "adres_2",
-            # postcode
-            "postcode": "postcode",
-            "postal code": "postcode",
-            "zip": "postcode",
-            "zip code": "postcode",
-            # gemeente / stad / city / plaats
-            "gemeente": "gemeente",
-            "stad": "gemeente",
-            "city": "gemeente",
-            "plaats": "gemeente",
-            "town": "gemeente",
-            # land
-            "land": "land",
-            "country": "land",
-            # btw (veel varianten)
-            "btw": "btw",
-            "vat": "btw",
-            "btw nummer": "btw",
-            "btw-nummer": "btw",
-            "btw number": "btw",
-            "vat number": "btw",
-            "btw number:": "btw",
-            "btw nr": "btw",
-            "btw nr.": "btw",
-            "btw-nr": "btw",
-            "btw-nr.": "btw",
-            "btw no": "btw",
-            "btw no.": "btw",
-            "vat no": "btw",
-            "vat no.": "btw",
-            "vat id": "btw",
-            "vat identification number": "btw",
-            "vat reg": "btw",
-            "vat reg.": "btw",
-            "vat reg number": "btw",
-            # contact
-            "contact sales": "contact_sales",
-            "contact_sales": "contact_sales",
-            "sales contact": "contact_sales",
-            "sales_contact": "contact_sales",
-            # email
-            "sales e-mail": "sales_email",
-            "sales email": "sales_email",
-            "e-mail sales": "sales_email",
-            "email sales": "sales_email",
-            "sales_email": "sales_email",
-            "email": "sales_email",
-            "mail": "sales_email",
-            # phone
-            "phone": "phone",
-            "phone number": "phone",
-            "telefoon": "phone",
-            "telefoon nummer": "phone",
-            "tel": "phone",
-            "tel. sales": "phone",
-            "tel sales": "phone",
-            # favorite
-            "favorite": "favorite",
-            "favoriet": "favorite",
-            "fav": "favorite",
-        }
-        norm = {}
-        for k, v in d.items():
-            lk = str(k).strip().lower()
-            if lk in key_map:
-                norm[key_map[lk]] = v
-
-        name = str(
-            norm.get("supplier")
-            or d.get("supplier")
-            or d.get("Leverancier")
-            or ""
-        ).strip()
-        if not name or name == "-":
-            raise ValueError("Supplier name is missing in record.")
-
-        fav = norm.get("favorite", d.get("favorite", False))
-        if isinstance(fav, str):
-            fav = fav.strip().lower() in ("1", "true", "yes", "y", "ja")
-
-        return Supplier(
-            supplier=name,
-            description=_to_str(norm.get("description")).strip() or None if ("description" in norm) else None,
-            supplier_id=_to_str(norm.get("supplier_id")).strip() or None if ("supplier_id" in norm) else None,
-            adres_1=_to_str(norm.get("adres_1")).strip() or None if ("adres_1" in norm) else None,
-            adres_2=_to_str(norm.get("adres_2")).strip() or None if ("adres_2" in norm) else None,
-            postcode=_to_str(norm.get("postcode")).strip() or None if ("postcode" in norm) else None,
-            gemeente=_to_str(norm.get("gemeente")).strip() or None if ("gemeente" in norm) else None,
-            land=_to_str(norm.get("land")).strip() or None if ("land" in norm) else None,
-            btw=_to_str(norm.get("btw")).strip() or None if ("btw" in norm) else None,
-            contact_sales=_to_str(norm.get("contact_sales")).strip() or None if ("contact_sales" in norm) else None,
-            sales_email=_to_str(norm.get("sales_email")).strip() or None if ("sales_email" in norm) else None,
-            phone=_to_str(norm.get("phone")).strip() or None if ("phone" in norm) else None,
-            favorite=bool(fav),
-        )
-
-
-class SuppliersDB:
-    def __init__(self, suppliers: List[Supplier]=None, defaults_by_production: Dict[str,str]=None):
-        self.suppliers: List[Supplier] = suppliers or []
-        self.defaults_by_production: Dict[str,str] = defaults_by_production or {}
-
-    @staticmethod
-    def load(path: str = SUPPLIERS_DB_FILE) -> "SuppliersDB":
-        if not os.path.exists(path):
-            return SuppliersDB()
-        try:
-            with open(path, "r", encoding="utf-8") as f:
-                data = json.load(f)
-            if isinstance(data, list):  # backward compat
-                sups = [Supplier(supplier=s) for s in data]
-                return SuppliersDB(sups, {})
-            sups_raw = data.get("suppliers", [])
-            sups = []
-            for rec in sups_raw:
-                try:
-                    sups.append(Supplier.from_any(rec))
-                except Exception:
-                    pass
-            defaults = data.get("defaults_by_production", {}) or {}
-            return SuppliersDB(sups, defaults)
-        except Exception:
-            return SuppliersDB()
-
-    def save(self, path: str = SUPPLIERS_DB_FILE) -> None:
-        data = {
-            "suppliers": [asdict(s) for s in self.suppliers],
-            "defaults_by_production": self.defaults_by_production
-        }
-        with open(path, "w", encoding="utf-8") as f:
-            json.dump(data, f, indent=2, ensure_ascii=False)
-
-    def suppliers_sorted(self) -> List[Supplier]:
-        return sorted(self.suppliers, key=lambda s: (not s.favorite, s.supplier.lower()))
-
-    def find(self, query: str) -> List[Supplier]:
-        q = (query or "").strip().lower()
-        if not q:
-            return self.suppliers_sorted()
-        L = []
-        for s in self.suppliers:
-            hay = " ".join([
-                s.supplier or "",
-                s.description or "",
-                s.supplier_id or "",
-                s.adres_1 or "",
-                s.adres_2 or "",
-                s.postcode or "",
-                s.gemeente or "",
-                s.land or "",
-                s.btw or "",
-                s.contact_sales or "",
-                s.sales_email or "",
-                s.phone or "",
-            ]).lower()
-            if q in hay:
-                L.append(s)
-        L.sort(key=lambda s: (not s.favorite, s.supplier.lower()))
-        return L
-
-    def display_name(self, s: Supplier) -> str:
-        return f"{'★ ' if s.favorite else ''}{s.supplier}"
-
-    def _idx_by_name(self, name: str) -> int:
-        for i, s in enumerate(self.suppliers):
-            if s.supplier.strip().lower() == str(name).strip().lower():
-                return i
-        return -1
-
-    def add(self, supplier: Supplier) -> bool:
-        if self._idx_by_name(supplier.supplier) >= 0:
-            return False
-        self.suppliers.append(supplier)
-        return True
-
-    def upsert(self, supplier: Supplier) -> None:
-        """Update bestaande met dezelfde naam, anders voeg toe."""
-        i = self._idx_by_name(supplier.supplier)
-        if i >= 0:
-            cur = self.suppliers[i]
-            for f in asdict(supplier):
-                val = getattr(supplier, f)
-                if val not in (None, ""):
-                    setattr(cur, f, val)
-        else:
-            self.suppliers.append(supplier)
-
-    def remove(self, name: str) -> bool:
-        i = self._idx_by_name(name)
-        if i >= 0:
-            self.suppliers.pop(i)
-            self.defaults_by_production = {k:v for k,v in self.defaults_by_production.items() if v.lower()!=name.lower()}
-            return True
-        return False
-
-    def clear_all(self):
-        self.suppliers = []
-        self.defaults_by_production = {}
-
-    def toggle_fav(self, name: str) -> bool:
-        i = self._idx_by_name(name)
-        if i >= 0:
-            self.suppliers[i].favorite = not self.suppliers[i].favorite
-            return True
-        return False
-
-    def set_default(self, production: str, supplier_name: str) -> None:
-        if self._idx_by_name(supplier_name) < 0:
-            raise ValueError(f"Supplier '{supplier_name}' not found")
-        self.defaults_by_production[str(production)] = supplier_name
-
-    def get_default(self, production: str) -> Optional[str]:
-        return self.defaults_by_production.get(str(production))
-
+from models import Supplier
+from suppliers_db import SuppliersDB, SUPPLIERS_DB_FILE
 
 # --------------------------- BOM & files ---------------------------
